adds support for tembo delete (for docker)
