+++ src/cmd/instance.rs
@@ -5,7 +5,6 @@ use simplelog::*;
 pub mod create;
 pub mod list;
 pub mod start;
-pub mod stop;
 
 // handles all instance command calls
 pub fn execute(args: &ArgMatches) -> Result<()> {
@@ -14,7 +13,6 @@ pub fn execute(args: &ArgMatches) -> Result<()> {
         Some(("create", sub_matches)) => create::execute(sub_matches),
         Some(("list", sub_matches)) => list::execute(sub_matches),
         Some(("start", sub_matches)) => start::execute(sub_matches),
-        Some(("stop", sub_matches)) => stop::execute(sub_matches),
         _ => unreachable!(),
     };
