+++ src/cmd/mod.rs
@@ -2,6 +2,7 @@ pub mod apply;
 pub mod auth;
 pub mod context;
 pub mod database;
+pub mod delete;
 pub mod extension;
 pub mod init;
 pub mod instance;
